+++ lib/mongo/app.ex
@@ -7,7 +7,6 @@ defmodule Mongo.App do
     children = [
       worker(Mongo.IdServer, []),
       worker(Mongo.PBKDF2Cache, []),
-      worker(:gen_event, [local: Mongo.Events]),
       supervisor(Registry, [:duplicate, :events_registry])
     ]