+++ test/mongo/session_test.exs
@@ -18,7 +18,7 @@ defmodule Mongo.SessionTest do
 
   @tag :mongo_3_6
   test "simple session insert", %{top: top} do
-    coll = unique_name()
+    coll = unique_collection()
 
     {:ok, session} = Session.start_session(top, :write, [])
     {:ok, %InsertOneResult{:inserted_id => id}} = Mongo.insert_one(top, coll, %{name: "Greta"}, session: session)
@@ -96,7 +96,7 @@ defmodule Mongo.SessionTest do
   @tag :mongo_3_6
   test "explicit_sessions", %{top: top} do
 
-    coll = unique_name()
+    coll = unique_collection()
 
     {:ok, session} = Session.start_session(top, :write, [])
     {:ok, %InsertOneResult{:inserted_id => id}} = Mongo.insert_one(top, coll, %{name: "Greta"}, session: session)
@@ -263,7 +263,7 @@ defmodule Mongo.SessionTest do
   @tag :mongo_4_2
   test "check unordered bulk with transaction", %{top: top} do
 
-    coll = unique_name()
+    coll = unique_collection()
     Mongo.insert_one(top, coll, %{name: "Wuff"})
     Mongo.delete_many(top, coll, %{})
 
@@ -291,7 +291,7 @@ defmodule Mongo.SessionTest do
   @tag :mongo_4_2
   test "check invalid unordered bulk with transaction", %{top: top} do
 
-    coll = "this-collection-does-not-exists"
+    coll = unique_collection()
 
     bulk = coll
            |> UnorderedBulk.new()
@@ -302,22 +302,26 @@ defmodule Mongo.SessionTest do
            |> UnorderedBulk.update_one(%{name: "Tom"}, %{"$set": %{kind: "dog"}})
            |> UnorderedBulk.update_one(%{name: "Waldo"}, %{"$set": %{kind: "dog"}})
 
-    #{:error, [result|_xs]} =
-      :error = Session.with_transaction(top, fn opts ->
+    cmd = [
+      configureFailPoint: "failCommand",
+      mode: [times: 1],
+      data: [errorCode: 3, failCommands: ["insert"]]
+    ]
+
+    assert {:ok, _} = Mongo.admin_command(top, cmd)
+
+    {:error, [result|_xs]} = Session.with_transaction(top, fn opts ->
 
       %BulkWriteResult{errors: errors} = result = BulkWrite.write(top, bulk, opts)
-
-      IO.puts inspect result
 
       case Enum.empty?(errors) do
          true  -> {:ok, result}
          false -> {:error, errors}
       end
 
-      :error
     end, w: 1)
 
-    # assert 263  == result["code"]
+    assert 3  == result.code
     assert {:ok, 0} == Mongo.count(top, coll, %{})
 
   end
@@ -325,7 +329,7 @@ defmodule Mongo.SessionTest do
   @tag :mongo_4_2
   test "check streaming bulk with transaction", %{top: top} do
 
-    coll = unique_name()
+    coll = unique_collection()
     Mongo.insert_one(top, coll, %{name: "Wuff"})
     Mongo.delete_many(top, coll, %{})
 
@@ -362,7 +366,7 @@ defmodule Mongo.SessionTest do
 
   @tag :mongo_4_2
   test "check ordered bulk with transaction", %{top: top} do
-    coll = unique_name()
+    coll = unique_collection()
     Mongo.insert_one(top, coll, %{name: "Wuff"})
     Mongo.delete_many(top, coll, %{})